conferencia_produtos/backend/requirements.txt
fastapi
uvicorn
firebirdsql
psycopg2
python-dotenv
redis
